corrected <archive_import_files>0<archive_import_files> not working
when setting the <import_options> <archive_import_files> flag to 0, the
file will still be archived. Casting the SimpleXMLElement to a bool
results in true no matter the value of the XML string the node
contains. Must cast to string first to extract the value of the node,
then cast to bool.